pizzarree_api/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.4
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
django-taggit==5.0.1
djangorestframework==3.14.0
Faker==20.1.0
phonenumbers==8.13.26
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
